+++ CHANGES.rst
@@ -6,7 +6,9 @@
 1.0.0a12 (unreleased)
 =====================
 
-- Nothing changed yet.
+- Add support for zope.schema 4.7.0 and nti.schema 1.5.0. Drop support
+  for older versions, which includes dropping support for
+  ``dm.zope.schema.Object`` fields.
 
 
 1.0.0a11 (2018-08-29)